# frontend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: MTN Bulk Data Wholesale Service API
  version: 0.1.0
  description: Backend contract required by the frontend-only MTN Bulk Data Wholesale Service console.
  x-api-requirements:
    - Every response must use the standard success/error envelope documented in components.schemas.ApiEnvelope and components.schemas.ErrorEnvelope.
    - Authenticated operations require the hardened HTTP-only JWT session cookie unless the operation explicitly sets security to an empty list.
    - Paginated table endpoints must return page, limit, total, totalPages, hasNextPage, and hasPreviousPage in meta.
    - Phone and MSISDN input must be normalized to E.164 Uganda numbers matching +256(77|78|79|76|39)XXXXXXX.
    - List endpoints with search parameters must support backend search because the frontend uses asynchronous API search when local table search has no matches.
    - Payment initiation must return a pending transaction and payment session that can be tracked globally until confirmation, expiry, or failure.
    - Customer account activation must follow the solution-document flow: activation link, email OTP verification, password creation, activation notification, then immediate portal access.
    - Bundle package validity is fixed at 30 days; configurable bundle fields are service code, name, volume up to 4 TB, price, status, and visibility.
...
servers:
  - url: https://backend.example.com/api
    description: Live backend base URL
security:
  - cookieAuth: []
...
paths:
...
  /provisioning/group-member:
    post:
      tags: [Provisioning]
      summary: Add Provisioning group member
      description: Privileged Provisioning operation that adds a secondary MSISDN under a primary subscription. Admin and support roles only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisioningAddGroupMemberRequest"
      responses:
        "200":
          description: Group member request accepted by Provisioning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningAddGroupMemberResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
  /provisioning/group-members/bulk:
    post:
      tags: [Provisioning]
      summary: Bulk add Provisioning group members
      description: Privileged Provisioning operation that adds multiple secondary MSISDNs in one provider request. Admin and support roles only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisioningAddGroupMembersBulkRequest"
      responses:
        "200":
          description: Bulk group-member request accepted by Provisioning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningAddGroupMembersBulkResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
  /provisioning/group-member/delete:
    post:
      tags: [Provisioning]
      summary: Delete Provisioning group member
      description: Privileged Provisioning operation that removes a secondary MSISDN from a grouped subscription. Admin and support roles only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisioningDeleteGroupMemberRequest"
      responses:
        "200":
          description: Group member delete request accepted by Provisioning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningDeleteGroupMemberResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
  /provisioning/subscriptions/update:
    post:
      tags: [Provisioning]
      summary: Update Provisioning subscription
      description: Privileged Provisioning operation used after successful payment to apply additional provisioning/top-up attempts. Admin and support roles only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisioningUpdateSubscriptionRequest"
      responses:
        "200":
          description: Subscription update request accepted by Provisioning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningUpdateSubscriptionResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
  /provisioning/subscriber:
    post:
      tags: [Provisioning]
      summary: Add Provisioning subscriber
      description: Privileged Provisioning operation that registers a primary subscriber. PCRF remains abstracted behind the backend adapter and is not exposed as a public API resource. Admin and support roles only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisioningAddSubscriberRequest"
            example:
              msisdn: "256779999707"
              transactionId: "78954566743"
      responses:
        "200":
          description: Subscriber request accepted by Provisioning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningAddSubscriberResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-mtn_bds_session
      description: HTTP-only, Secure, SameSite=Strict signed JWT session cookie. Local development may use mtn_bds_session over http.
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            validationError:
              $ref: "#/components/examples/ErrorResponseExample"
  examples:
...
    ErrorResponseExample:
      summary: Validation error
      value:
        success: false
        message: Validation failed
        data: null
  schemas:
    ApiEnvelope:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
        message:
          type: string
        data: {}
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    PaginationMeta:
      type: object
      required: [page, limit, total, totalPages, hasNextPage, hasPreviousPage]
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 1
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean
    ErrorEnvelope:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
        data:
          type: "null"
...
    ProvisioningNormalizedMsisdn:
      type: string
      description: Provisioning-normalized Uganda MTN MSISDN without the leading plus sign.
      pattern: "^256(77|78|79|76|39)\\d{7}$"
      example: "256772990001"
    ProvisioningGroupMemberPair:
      type: object
      required: [secondaryMsisdn, primaryMsisdn]
      properties:
        secondaryMsisdn:
          $ref: "#/components/schemas/ProvisioningNormalizedMsisdn"
        primaryMsisdn:
          $ref: "#/components/schemas/ProvisioningNormalizedMsisdn"
    ProvisioningAddGroupMemberRequest:
      $ref: "#/components/schemas/ProvisioningGroupMemberPair"
    ProvisioningAddGroupMembersBulkRequest:
      type: object
      required: [groupMembers]
      properties:
        groupMembers:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: "#/components/schemas/ProvisioningGroupMemberPair"
    ProvisioningDeleteGroupMemberRequest:
      type: object
      required: [secondaryMsisdn]
      properties:
        secondaryMsisdn:
          $ref: "#/components/schemas/ProvisioningNormalizedMsisdn"
    ProvisioningUpdateSubscriptionRequest:
      type: object
      required:
        - primaryMsisdn
        - serviceCode
        - transactionId
        - topupValue
        - updateAttemptCount
      properties:
        primaryMsisdn:
          $ref: "#/components/schemas/ProvisioningNormalizedMsisdn"
        serviceCode:
          type: string
          example: DATA_BUNDLE_CODE
        transactionId:
          type: string
          example: "567123456"
        topupValue:
          type: integer
          minimum: 1
          description: Top-up volume in kilobytes (KB). Purchase workflows must convert bundle volume to KB before calling this endpoint.
          example: 1024
        updateAttemptCount:
          type: integer
          minimum: 0
          example: 4
    ProvisioningAddSubscriberRequest:
      type: object
      required: [msisdn, transactionId]
      properties:
        msisdn:
          $ref: "#/components/schemas/ProvisioningNormalizedMsisdn"
        transactionId:
          type: string
          example: "78954566743"
    ProvisioningCommandResult:
      type: object
      required:
        [
          requestId,
          operation,
          accepted,
          processedAt,
          providerStatusCode,
          request,
        ]
      properties:
        requestId:
          type: string
          format: uuid
        operation:
          type: string
          enum:
            - add_group_member
            - add_group_members_bulk
            - delete_group_member
            - update_subscription
            - add_subscriber
            - subscribe_service
        accepted:
          type: boolean
          example: true
        processedAt:
          type: string
          format: date-time
        providerStatusCode:
          type: integer
          example: 200
        request:
          type: object
        providerResponse:
          type: object
          additionalProperties: true
          nullable: true
          example:
            successful: true
            resultCode: "0"
            message: Operation succeeded
    ProvisioningAddGroupMemberResult:
      allOf:
        - $ref: "#/components/schemas/ProvisioningCommandResult"
        - type: object
          properties:
            request:
              $ref: "#/components/schemas/ProvisioningAddGroupMemberRequest"
    ProvisioningAddGroupMembersBulkResult:
      allOf:
        - $ref: "#/components/schemas/ProvisioningCommandResult"
        - type: object
          properties:
            request:
              $ref: "#/components/schemas/ProvisioningAddGroupMembersBulkRequest"
    ProvisioningDeleteGroupMemberResult:
      allOf:
        - $ref: "#/components/schemas/ProvisioningCommandResult"
        - type: object
          properties:
            request:
              $ref: "#/components/schemas/ProvisioningDeleteGroupMemberRequest"
    ProvisioningUpdateSubscriptionResult:
      allOf:
        - $ref: "#/components/schemas/ProvisioningCommandResult"
        - type: object
          properties:
            request:
              $ref: "#/components/schemas/ProvisioningUpdateSubscriptionRequest"
    ProvisioningAddSubscriberResult:
      allOf:
        - $ref: "#/components/schemas/ProvisioningCommandResult"
        - type: object
          properties:
            request:
              $ref: "#/components/schemas/ProvisioningAddSubscriberRequest"
...
    ProvisioningAddGroupMemberResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ProvisioningAddGroupMemberResult"
    ProvisioningAddGroupMembersBulkResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ProvisioningAddGroupMembersBulkResult"
    ProvisioningDeleteGroupMemberResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ProvisioningDeleteGroupMemberResult"
    ProvisioningUpdateSubscriptionResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ProvisioningUpdateSubscriptionResult"
    ProvisioningAddSubscriberResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ProvisioningAddSubscriberResult"
...